Fix: 실시간 인기 질문 컴포넌트 분리
